Move rest of editor logic to editor.py
